feat: 1537 poc add swagger docs and ui
