+++ package.json
@@ -54,7 +54,6 @@
     "@types/easy-soap-request": "^4.1.4",
     "@types/swagger-jsdoc": "^6.0.4",
     "@types/swagger-ui": "^3.52.4",
-    "apidoc": "^1.2.0",
     "axios": "^1.6.5",
     "dedent": "^1.5.1",
     "dotenv": "^16.3.2",
@@ -66,7 +65,6 @@
     "koa": "^2.15.0",
     "koa-bodyparser": "^4.4.1",
     "koa-pino-logger": "^4.0.0",
-    "koa-swagger-decorator": "^1.8.7",
     "koa2-swagger-ui": "^5.10.0",
     "mssql": "^9.3.2",
     "onecore-types": "^1.21.1",
@@ -80,7 +78,6 @@
     "randexp": "^0.5.3",
     "strong-soap": "^4.0.0",
     "swagger-jsdoc": "^6.2.8",
-    "swagger-ui-koa": "^0.0.1",
     "tedious": "^16.6.1",
     "zod": "^3.23.8"
   }
