+++ src/api.ts
@@ -4,6 +4,7 @@ import { routes as leaseRoutes } from './services/lease-service'
 import { routes as creditSafeRoutes } from './services/creditsafe'
 import { routes as healthRoutes } from './services/health-service'
 import { routes as applicantsRoutes } from './services/lease-service/routes/applicants'
+import { routes as swagggerRoutes } from './services/swagger'
 
 const router = new KoaRouter()
 
@@ -11,5 +12,6 @@ leaseRoutes(router)
 creditSafeRoutes(router)
 healthRoutes(router)
 applicantsRoutes(router)
+swagggerRoutes(router)
 
 export default router
