+++ src/app.ts
@@ -9,44 +9,8 @@ import { logger, loggerMiddlewares } from 'onecore-utilities'
 const app = new Koa()
 
 import { koaSwagger } from 'koa2-swagger-ui'
-import { SwaggerRouter } from 'koa-swagger-decorator'
-const swaggerRouter = new SwaggerRouter()
 
 app.use(cors())
-
-//const swaggerJsdoc = import('swagger-jsdoc');
-import swaggerJsdoc from 'swagger-jsdoc'
-
-const options = {
-  definition: {
-    openapi: '3.0.0',
-    info: {
-      title: 'Hello World',
-      version: '1.0.0',
-    },
-  },
-  apis: ['./src/services/lease-service/routes/*.ts'],
-  servers: [
-    {
-      url: 'http://localhost:3000',
-    },
-  ],
-}
-
-const swaggerSpec = swaggerJsdoc(options)
-
-console.log('spec: ')
-console.log(swaggerSpec)
-
-// app.use(
-//   koaSwagger({
-//     routePrefix: '/swagger',
-//     swaggerOptions: {
-//       url: 'http://petstore.swagger.io/v2/swagger.json', // example path to json
-//       //spec: swaggerSpec,
-//     },
-//   })
-// )
 
 app.use(
   koaSwagger({
@@ -56,16 +20,6 @@ app.use(
     },
   })
 )
-
-/*swaggerRouter.swagger({
-  title: 'API Documentation',
-  description: 'API Documentation for Koa TypeScript API',
-  version: '1.0.0',
-})
-
-swaggerRouter.mapDir('src/services/lease-service/routes')
-
-app.use(swaggerRouter.routes()).use(swaggerRouter.allowedMethods())*/
 
 app.on('error', (err) => {
   logger.error(err)
